backend/requirements.txt
flask
flask-cors
google-genai
gunicorn
google-cloud-firestore
